requirements.txt
nextcord==2.3.2
python-dotenv==0.21.0
requests==2.28.1
spoonacular==3.0
